Added Pump functionality to fill first car in queue
Added fillFirstVehicleTank method in Pump, will fill the tank of the
vehicle in index 0 of the pumps queue. I don't thin the 'currentVehicle'
field is needed, as that should always be the first vehicle in the
queue. plus, this would make complications in calculating how much space
is elft int he queue, as 'currentVehicle' should still be in the queue